refactor: extract remaining constructions to per-type packages
Moves all remaining construction classes from Constructions.ts to
their element-type packages:

- polygon/PolygonConstructions.ts — 18 classes (incl. PolyConstruction)
- sector/SectorConstructions.ts — 4 classes
- plane/PlaneConstructions.ts — 5 classes
- sphere/SphereConstructions.ts — 2 classes
- polyhedron/PolyhedronConstructions.ts — 4 classes

Constructions.ts: 2036 → 450 lines. Now contains only:
- Enum definitions (PointConstructions, LineConstructions, etc.)
- Abstract Construction base class with validateSignature()
- Imports from all 8 per-type construction files
- The constructions registration array

Build passes, 98 tests pass.
